requirements.txt
openai
google-cloud-speech
requests
pyaudio
python-dotenv